Revert "move test_dpsk_v32_encoding_path back to GPU suite"
This reverts commit e6de64a96ee5ffa721667d9019b4137fe44161a5.
